refactor: move the Prometheus client out from under internal/api
Neither half of the package is a REST concern. Client runs instant and
range queries and is the concrete type behind internal/mcp's
MetricsQuerier, which main wires; Proxy reaches back into the transport
for error rendering through an ErrorWriter rather than importing it, which
is exactly what lets the package stand alone. Nesting it under api/ was
historical and told the next reader it belonged to one transport.

internal/prom keeps the three result types, so a caller that only names a
series — internal/recommendations, internal/mcp — still does not pull in
an HTTP client to do it. Adds a doc.go saying so.
