+++ internal/api/router.go
@@ -10,12 +10,12 @@ import (
 	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/api/types/swarm"
 
-	"github.com/radiergummi/cetacean/internal/api/prometheus"
 	"github.com/radiergummi/cetacean/internal/api/sse"
 	"github.com/radiergummi/cetacean/internal/auth"
 	"github.com/radiergummi/cetacean/internal/cache"
 	"github.com/radiergummi/cetacean/internal/config"
 	"github.com/radiergummi/cetacean/internal/metrics"
+	"github.com/radiergummi/cetacean/internal/prometheus"
 )
 
 // Resyncer triggers a manual full re-fetch of cluster state. Backed by the
